svh/requirements.txt
requests
django
uwsgi
ipython
psycopg2-binary
imohash
pillow
opencv-python
factory_boy
django-mptt
celery
moviepy
pyyaml
qbittorrent-api
ffmpeg-python
pika